Fix type conversion warnings.

+++ src/platform/headless.cpp
@@ -120,11 +120,12 @@ void PaintGraphics() {
     pixmap->format = Pixmap::Format::BGRA;
     pixmap->width  = camera.width;
     pixmap->height = camera.height;
-    pixmap->stride = cairo_format_stride_for_width(CAIRO_FORMAT_RGB24, camera.width);
+    pixmap->stride = cairo_format_stride_for_width(CAIRO_FORMAT_RGB24, (int)camera.width);
     pixmap->data   = std::vector<uint8_t>(pixmap->stride * pixmap->height);
     cairo_surface_t *surface =
         cairo_image_surface_create_for_data(&pixmap->data[0], CAIRO_FORMAT_RGB24,
-                                            pixmap->width, pixmap->height, pixmap->stride);
+                                            (int)pixmap->width, (int)pixmap->height,
+                                            (int)pixmap->stride);
     cairo_t *context = cairo_create(surface);
 
     CairoRenderer canvas;
